requirements.txt
matplotlib
pyqt5
pyqtgraph
scipy
pytorch
tqdm
wandb